requirements.txt
mysql-connector-python==8.0.33
tkcalendar==1.6.1
babel==2.17.0
pip==25.0.1